rename directory 'entities' to 'entity'
