--- src/apps/user/dtos/__init__.py
@@ -1,17 +1,0 @@
-from .dtos import (
-    AdminCreateUserDTO,
-    AdminPasswordUpdateDTO,
-    AuthDTO,
-    UserCreateDTO,
-    UserPasswordUpdateDTO,
-    UserUpdateDTO,
-)
-
-__all__ = [
-    'UserCreateDTO',
-    'UserUpdateDTO',
-    'AuthDTO',
-    'AdminCreateUserDTO',
-    'UserPasswordUpdateDTO',
-    'AdminPasswordUpdateDTO',
-]
