+++ src/apps/user/services/authorize_user_service.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from src.apps.user.domain.entities import User
+from src.apps.user.domain.entity import User
 from src.apps.user.exceptions import (
     UserNotAdminError,
     UserNotFoundError,
