+++ src/apps/user/services/get_user_service.py
@@ -1,6 +1,6 @@
 from typing import List, Optional
 
-from src.apps.user.domain.entities import User
+from src.apps.user.domain.entity import User
 from src.apps.user.exceptions import TokenExpiredError, UserNotFoundError
 from src.apps.user.protocols import JWTServiceProtocol
 from src.apps.user.repositories import IUserRepository
