+++ src/apps/user/services/update_user_service.py
@@ -1,4 +1,4 @@
-from src.apps.user.domain.entities import User
+from src.apps.user.domain.entity import User
 from src.apps.user.dtos import UserUpdateDTO
 from src.apps.user.exceptions import UserAlreadyExistsError
 from src.apps.user.protocols import JWTServiceProtocol
