+++ src/data_access/mappers/user_mapper.py
@@ -1,4 +1,4 @@
-from src.apps.user.domain.entities import User
+from src.apps.user.domain.entity import User
 from src.data_access.models import UserModel
